renaming the components to more appropriate names and formatting
