+++ src/components/RestrictedContent.jsx
@@ -1,0 +1,22 @@
+import React from "react";
+import { connect } from "react-redux";
+
+const RestrictedContent = props => {
+  let article = props.singleArticle;
+
+  let shortContent = article.content.substring(0, 20) + "...";
+  return (
+    <>
+      <p>{shortContent}</p>
+      <button>Buy Subscription</button>
+    </>
+  );
+};
+
+const mapStateToProps = state => {
+  return {
+    singleArticle: state.singleArticle
+  };
+};
+
+export default connect(mapStateToProps)(RestrictedContent);
